part1/requirements.txt
aiohttp
asyncio
google-cloud-pubsub